added ExceptionHandlingMiddleware, added a custom NotFoundException handler

+++ src/OnlineStore.CatalogService/OnlineStore.CatalogService.Application/Handlers/DeleteProductByIdHandler.cs
@@ -1,4 +1,5 @@
 ﻿using MediatR;
+using OnlineStore.CatalogService.Application.Exceptions;
 using OnlineStore.CatalogService.Application.Queries;
 using OnlineStore.CatalogService.Domain.Repositories;
 
@@ -15,6 +16,11 @@ namespace OnlineStore.CatalogService.Application.Handlers
 
         public async Task<bool> Handle(DeleteProductByIdQuery request, CancellationToken cancellationToken)
         {
+            if (request.Id == null)
+            {
+                throw new NotFoundException("Product not found");
+            }
+
             return await _productRepository.DeleteProduct(request.Id);
         }
     }
